Make bulk edit form function component
